backend_flask_new/requirements.txt
argon2-cffi
flask
flask-jwt-extended
marshmallow-sqlalchemy
python-dotenv
sqlalchemy